Backend/scripts/requirements.txt
sqlalchemy
pandas
psycopg2-binary
google-cloud-storage
google-cloud-aiplatform
